renamed _spatialdata.py to spatialdata.py; using src/
