+++ .flake8
@@ -47,7 +47,7 @@ per-file-ignores =
     tests/*: D
     */__init__.py: F401
     # Ignore for compat checks at top of file
-    spatialdata/__init__.py: E402
+    src/spatialdata/__init__.py: E402
     # TODO: this file is weird, but it's not actually causing a bug and fixing it is
     # signifigantly harder than ignoring it
-    spatialdata/_constants/_enum.py: B024
+    src/spatialdata/_constants/_enum.py: B024
