+++ .pre-commit-config.yaml
@@ -29,7 +29,7 @@ repos:
       hooks:
           - id: mypy
             additional_dependencies: [numpy==1.24, types-requests]
-            exclude: tests/|docs/|temp/|spatialdata/_core/reader.py
+            exclude: tests/|docs/|temp/
     - repo: https://github.com/asottile/yesqa
       rev: v1.4.0
       hooks:
