--- spatialdata/_core/__init__.py
@@ -1,1 +1,0 @@
-from spatialdata._core._spatialdata import SpatialData
