+++ src/spatialdata/_compat.py
