+++ src/spatialdata/_constants/_constants.py
