+++ src/spatialdata/_constants/_enum.py
