+++ src/spatialdata/_core/__init__.py
