+++ src/spatialdata/_core/_rasterize.py
