+++ src/spatialdata/_core/core_utils.py
