+++ src/spatialdata/_core/data_extent.py
