+++ src/spatialdata/_core/models.py
