+++ src/spatialdata/_core/ngff/ngff_coordinate_system.py
