+++ src/spatialdata/_core/ngff/ngff_transformations.py
