+++ src/spatialdata/_core/spatialdata.py
