+++ src/spatialdata/_dataloader/datasets.py
