+++ src/spatialdata/_io/_utils.py
