+++ src/spatialdata/_io/format.py
