+++ src/spatialdata/_io/write.py
