+++ src/spatialdata/_types.py
