+++ src/spatialdata/utils.py
